refactor: Use react-query
