+++ apps/web/src/hooks/useBNBPrice.ts
@@ -5,14 +5,16 @@ import { chainlinkOracleABI } from 'config/abi/chainlinkOracle'
 import contracts from 'config/constants/contracts'
 import { publicClient } from 'utils/wagmi'
 import { formatUnits } from 'viem'
-import useSWR from 'swr'
 import { FAST_INTERVAL } from 'config/constants'
+import { useQuery } from '@tanstack/react-query'
 
 // for migration to bignumber.js to avoid breaking changes
-export const useBNBPriceAsBN = () => {
-  const { data } = useSWR(['bnbPriceAsBN'], async () => new BigNumber(await getBNBPriceFromOracle()), {
-    dedupingInterval: FAST_INTERVAL,
-    refreshInterval: FAST_INTERVAL,
+export const useBNBPrice = () => {
+  const { data } = useQuery<BigNumber, Error>({
+    queryKey: ['bnbPrice'],
+    queryFn: async () => new BigNumber(await getBNBPriceFromOracle()),
+    staleTime: FAST_INTERVAL,
+    refetchInterval: FAST_INTERVAL,
   })
   return data ?? BIG_ZERO
 }
@@ -23,8 +25,6 @@ export const getBNBPriceFromOracle = async () => {
     address: contracts.chainlinkOracleBNB[ChainId.BSC],
     functionName: 'latestAnswer',
   })
-
-  console.info(data)
 
   return formatUnits(data, 8)
 }
