+++ apps/web/src/views/Nft/market/components/BuySellModals/SellModal/SetPriceStage.tsx
@@ -4,7 +4,7 @@ import { escapeRegExp } from 'utils'
 import { useTranslation } from '@pancakeswap/localization'
 import { NftToken } from 'state/nftMarket/types'
 import { useGetCollection } from 'state/nftMarket/hooks'
-import { useBNBPriceAsBN } from 'hooks/useBNBPriceAsBN'
+import { useBNBPrice } from 'hooks/useBNBPrice'
 import { Divider } from '../shared/styles'
 import { GreyedOutContainer, BnbAmountCell, RightAlignedInput, FeeAmountCell } from './styles'
 
@@ -42,9 +42,9 @@ const SetPriceStage: React.FC<React.PropsWithChildren<SetPriceStageProps>> = ({
   const { creatorFee = '', tradingFee = '' } = useGetCollection(nftToSell.collectionAddress) || {}
   const creatorFeeAsNumber = parseFloat(creatorFee)
   const tradingFeeAsNumber = parseFloat(tradingFee)
-  const bnbPrice = useBNBPriceAsBN()
+  const bnbBusdPrice = useBNBPrice()
   const priceAsFloat = parseFloat(price)
-  const priceInUsd = bnbPrice.multipliedBy(priceAsFloat).toNumber()
+  const priceInUsd = bnbBusdPrice.multipliedBy(priceAsFloat).toNumber()
 
   const priceIsOutOfRange = priceAsFloat > MAX_PRICE || priceAsFloat < MIN_PRICE
