+++ apps/web/src/views/Nft/market/components/CollectibleCard/CardBody.tsx
@@ -1,7 +1,7 @@
 import { Box, CardBody, Flex, Text } from '@pancakeswap/uikit'
 import { useTranslation } from '@pancakeswap/localization'
 import { isAddress } from 'utils'
-import { useBNBPriceAsBN } from 'hooks/useBNBPriceAsBN'
+import { useBNBPrice } from 'hooks/useBNBPrice'
 import PreviewImage from './PreviewImage'
 import { CostLabel, LowestPriceMetaRow, MetaRow } from './styles'
 import LocationTag from './LocationTag'
@@ -18,7 +18,7 @@ const CollectibleCardBody: React.FC<React.PropsWithChildren<CollectibleCardProps
 }) => {
   const { t } = useTranslation()
   const { name } = nft
-  const bnbBusdPrice = useBNBPriceAsBN()
+  const bnbBusdPrice = useBNBPrice()
   const isPancakeBunny = isAddress(nft.collectionAddress) === pancakeBunniesAddress
   const { isFetching, lowestPrice } = useGetLowestPriceFromNft(nft)
